[fix] Extract task_manager into a shared parallel_query module
task_manager and task_execution_guard were defined in the hash_join
namespace, but the parallel merge join reused them via aliases - making
merge join appear to depend on hash join scaffolding. Move both classes
into a neutral px_task_manager.{hpp,cpp} under parallel_query so hash
join and merge join reference the same worker-coordination primitive.

- Behavior unchanged; the classes are relocated verbatim.
- hash_join keeps working through using-declarations, so px_hash_join.cpp
  needs no changes.
- px_merge_join_task.hpp no longer includes the hash join task header.
- task_execution_guard still references hash_join::spawn_manager (the
  accepted Tier-1 limitation); a spawn_manager-free guard can be split
  out later if a third parallel operator needs one.
